0.1.5.2 Beta
- Новая текстура сундука
- Новая текстура досок
- Фикс крафтов
- Фикс интерфейса паузы
- Поломан интерфейс крафтинга
- Ресайз окна работает

+++ core/src/com/kgc/sauw/UI/Elements/Image.java
@@ -10,7 +10,7 @@ public class Image extends InterfaceElement {
     private Texture texture;
     private TextureRegion textureRegion;
 
-    public Image(int x, int y, int w, int h) {
+    public Image(float x, float y, float w, float h) {
         setPosition(x, y);
         setSize(w, h);
     }
